Moved ServiceDefinition
